1a: Beginning Student Language - Expressions pt1
Exercise from 1a: Beginning Student Language
Practises use of Dr. Racket that uses BSL (Beginning Student Language; a more introductory type of language with the focus being learning the systematic program design method)
This exercise asks to create an expression in Dr. Racket to evaluate the value of the hyponese where the sides of the triangle are 3 and 4