My answer to exercises Module 3b
Problems:
Movie Starter - Design a compound data definition to store movie name, budget, and year released. Then create a function that takes 2 movies and gives the name of the movie that was more recently released
Student Starter - Design a compound data definition to store student name, grade, and if they have allergies (boolean). Then create a function that takes in a student and determines if the student should be placed on a special list (based on if grade is less than or equal to 6 and if the student has allergies)