Update WIP: Spinning Bears; Finished Part1
Finished the first part of the problem:
have a rotating bear appear in the middle of the screen when started. The rotating bear should be replaced with a new bear any time the mouse is clicked. 
The position of the new bear is where the mouse click event happened (x and y coordinates of mouse click)