Completed My Answer - Exercise Spinning Bears Pt2
Finished exercise including part 1 and part 2.
Part 2 builds off of Part 1. The result is a screen that starts off empty (run program with (main empty) in the terminal), then any time the mouse is clicked on the screen a new rotating bear is drawn on the screen. 
All rotating bears are saved in a list, therefore the number of rotating bears drawn on the screen should increase each time the mouse is clicked. 
No other inputs are accepted (i.e. if keyboard button is pressed then nothing happens)