My solution to the exercise: Making Rain Filtered
Exercise Problem:
Design a world program to make it rain where you want it to. In this version of the program you should remove drops that have already fallen off the screen from the world state. Be sure to use all appropriate helper function rules. There are a total of 9 functions in our solution, so this problem takes some time.

My solution uses 7 functions but it still completes the problem. Where the professor gets the other 2 functions was by:
 * having the function to draw the whole array that ends up calling another function to draw a single drop. Whereas I have a function that has recursive and draws the whole array
 * creating a function to store the new filtered array. Whereas I have the filtered array as a return as part of my tick (update y postion) function.

I can see some scenarios in which drawing each drop one by one may be handy but for this small program I prefer drawing the whole array.
However, for the other function that specifically returns a filtered array, I think that's more efficient as it's more granular and breaks the problem apart even more. My solution does still work but it could be improved on to be more efficient