My answer to exercise: BST P4 - Insert
Exercise problem:
Design a function that consumes an Integer, String and BST, and adds a node that has the given key and value to the tree. The node should be inserted in the proper place in the tree. The function can assume there is not already an entry for that number in the tree. The function should produce the new BST.

My solution: although the problem says we can assume that there is not already an entry for the key we want to input, I provided a test-case to check if the tree already has the same entry as the parameter. I incorporated this into my function solution so that if there is an existing entry then the function would return the same BST. 
Another test case was made to check if the input key is 0 (zero), if so then it will also just return the same BST.