FINAL My Answer: Exercise Render BST With Lines
Exercise Problem:
Design a function that renders a simple BST with lines. Note that the tree will not be given in advance, so recursion calls are necessary

My solution takes into account whether the next node is empty or not. If it is empty then the left/right branch won't be drawn. In some scenarios where the branches are very far apart, this causes the line to not stretch all the way to the node.

In comparison, although the professor's solution has cleaner lines that stretch, the downside is that their solution adds left AND right lines to all the nodes, regardless of whether or not the following node is empty. This can make the tree appear a bit confusing when the bottom row of the tree has branches but no key/values after it.

Overall, I prefer my solution as it's much clearer to understand the rendering. Improvements could be made to make the lines more aesthetically pleasing.